fix(webapp): rename records
