+++ www/webapp/src/components/Field/RecordHTTPS.vue
@@ -1,5 +1,5 @@
 <script>
-import RecordSVCB from './SVCB.vue';
+import RecordSVCB from './RecordSVCB.vue';
 
 export default {
   name: 'RecordHTTPS',
