+++ www/webapp/src/components/Field/RecordMX.vue
@@ -1,22 +1,26 @@
 <script>
 import { helpers, integer, between } from 'vuelidate/lib/validators';
-import RecordItem from '../RecordItem.vue';
+import RecordItem from './RecordItem.vue';
 
-// Allow for root label only
-const hostname = helpers.regex('hostname', /^((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))?[.]?$/);
+const hostname = helpers.regex('hostname', /^((([a-zA-Z0-9-]+\.?)+)|\.)$/);
 const trailingDot = helpers.regex('trailingDot', /[.]$/);
 
 const MAX16 = 65535;
 const int16 = between(0, MAX16);
 
 export default {
-  name: 'RecordSVCB',
+  name: 'RecordMX',
   extends: RecordItem,
   data: () => ({
     fields: [
-      { label: 'Priority', validations: { integer, int16 } },
-      { label: 'Target Name', validations: { hostname, trailingDot } },
-      { label: 'Service Parameters', validations: { }, optional: true },
+      {
+        label: 'Preference',
+        validations: { integer, int16 },
+      },
+      {
+        label: 'Hostname',
+        validations: { hostname, trailingDot },
+      },
     ],
     errors: {
       integer: 'Please enter an integer.',
